+++ src/main/java/com/google/bamboo/soy/format/SoyCodeStyleSettingsProvider.java
@@ -1,0 +1,51 @@
+// Copyright 2017 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.bamboo.soy.format;
+
+import com.google.bamboo.soy.SoyLanguage;
+import com.intellij.application.options.CodeStyleAbstractConfigurable;
+import com.intellij.application.options.CodeStyleAbstractPanel;
+import com.intellij.application.options.TabbedLanguageCodeStylePanel;
+import com.intellij.lang.Language;
+import com.intellij.openapi.options.Configurable;
+import com.intellij.psi.codeStyle.CodeStyleSettings;
+import com.intellij.psi.codeStyle.CodeStyleSettingsProvider;
+import org.jetbrains.annotations.NotNull;
+
+public class SoyCodeStyleSettingsProvider extends CodeStyleSettingsProvider {
+
+  @Override
+  public Language getLanguage() {
+    return SoyLanguage.INSTANCE;
+  }
+
+  @NotNull
+  @Override
+  public Configurable createSettingsPage(CodeStyleSettings settings,
+      CodeStyleSettings originalSettings) {
+    return new CodeStyleAbstractConfigurable(settings, originalSettings,
+        getConfigurableDisplayName()) {
+      @Override
+      protected CodeStyleAbstractPanel createPanel(CodeStyleSettings settings) {
+        return new TabbedLanguageCodeStylePanel(getLanguage(), settings, originalSettings) {
+          @Override
+          protected void initTabs(CodeStyleSettings settings) {
+            addIndentOptionsTab(settings);
+          }
+        };
+      }
+    };
+  }
+}
