+++ src/main/resources/codeSamples/IndentSettings.soy
@@ -1,0 +1,12 @@
+{namespace
+site.page.home}
+
+{template .site}
+  <div class="wrapper">
+    {call .page}
+      {param content kind="html"}
+          <div>Welcome to Bamboo Soy page.</div>
+      {/param}
+    {/call}
+  </div>
+{/template}
